+++ lib/process_hub/strategy/migration/hot_swap.ex
@@ -39,6 +39,13 @@ defmodule ProcessHub.Strategy.Migration.HotSwap do
   >   end
   > end
   > ```
+
+  > #### State Handover with Replication {: .warning}
+  >
+  > State handover is **not supported** when using the `ProcessHub.Strategy.Redundancy.Replication`
+  > strategy. With replication, multiple instances of a process run across the cluster, making
+  > state handover semantics undefined. If you attempt to use `handover: true` with replication,
+  > the hub will fail to start with `{:error, {:invalid_config, :handover_with_replication_not_supported}}`.
   """
 
   require Logger